requirements.txt
gradio==4.36.1
numpy==1.23.5
pandas==1.5.3
tensorflow-cpu==2.11.0
h5py==3.8.0